requirements.txt
aiofiles==23.2.1
fastapi==0.110.0
Jinja2==3.1.3
openai==1.12.0
python-dotenv==1.0.1
python-multipart==0.0.9
uvicorn==0.27.1
websockets==12.0
anthropic>=0.16.0
redis==5.2.1
boto3>=1.28.0
numpy